Add message authorization

+++ my-chat-backend/controllers/messages.js
@@ -1,5 +1,16 @@
 const messagesRouter = require('express').Router()
 const Message = require('../models/message')
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
+const User = require('../models/user')
+
+const getTokenFrom = req => {
+  const authorization = req.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 messagesRouter.get('/', (req, res, next) => {
   Message.find({})
@@ -21,35 +32,50 @@ messagesRouter.get('/:id', (req,res, next) => {
     .catch(next)
 })
 
-messagesRouter.delete('/:id', (req,res, next) => {
-  Message.findByIdAndDelete(req.params.id)
-    .then(() => {
-      res.status(204).end()
-    })
-    .catch(next)
-})
-
-messagesRouter.post('/', (req,res, next) => {
-  const body = req.body
-
-  if (!body.name || !body.content || body.name.trim() === '' || body.content.trim() === '') {
-    return res.status(400).json({
-      error: 'name or content missing'
+messagesRouter.delete('/:id', async (req,res, next) => {
+  const decodedToken = jwt.verify(getTokenFrom(req), config.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({
+      error: 'token invalid'
     })
   }
 
-  const message = new Message ({
-    name: body.name.trim(),
-    content: body.content.trim(),
+  const message = await Message.findById(req.params.id)
+  if (!message) {
+    return res.status(404).end()
+  }
+
+  if (message.user.toString() !== decodedToken.id) {
+    return res.status(403).json({
+      error: 'not allowed to delete this message'
+    })
+  }
+
+  await Message.findByIdAndDelete(req.params.id)
+    
+  await User.findByIdAndUpdate(decodedToken.id, {
+    $pull: { messages: message._id }
   })
 
-  message.save()
-    .then(savedMessage => {res.status(201).json(savedMessage)})
-    .catch(next)
+  res.status(204).end()
 })
 
-messagesRouter.put('/:id', (req,res, next) => {
+messagesRouter.post('/', async (req,res, next) => {
   const body = req.body
+
+  const decodedToken = jwt.verify(getTokenFrom(req), config.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ 
+      error: 'token invalid' 
+    })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return res.status(401).json({ 
+      error: 'UserId missing or invalid' 
+    })
+  }
 
   if (!body.content || body.content.trim() === '') {
     return res.status(400).json({
@@ -57,18 +83,51 @@ messagesRouter.put('/:id', (req,res, next) => {
     })
   }
 
-  Message.findByIdAndUpdate(req.params.id,
-    { content: body.content },
-    { returnDocument: 'after', runValidators: true }
-  )
-    .then(updatedMessage => {
-      if (updatedMessage) {
-        res.json(updatedMessage)
-      } else {
-        res.status(404).end()
-      }
+  const message = new Message ({
+    name: user.name || user.username,
+    content: body.content.trim(),
+    user: user._id,
+  })
+
+  const savedMessage = await message.save()
+
+  user.messages = user.messages.concat(savedMessage._id)
+  await user.save()
+
+  res.status(201).json(savedMessage)
+})
+
+messagesRouter.put('/:id', async (req,res, next) => {
+  const body = req.body
+
+  const decodedToken = jwt.verify(getTokenFrom(req), config.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({
+      error: 'token invalid'
     })
-    .catch(next)
+  }
+
+  const message = await Message.findById(req.params.id)
+  if (!message) {
+    return res.status(404).end()
+  }
+
+  if (message.user.toString() !== decodedToken.id) {
+    return res.status(403).json({
+      error: 'not allowed to edit this message'
+    })
+  }
+
+  if (!body.content || body.content.trim() === '') {
+    return res.status(400).json({
+      error: 'content missing'
+    })
+  }
+  
+  message.content = body.content.trim()
+
+  const updatedMessage = await message.save()
+  res.status(200).json(updatedMessage)
 }) 
 
 module.exports = messagesRouter
